♻️ Refactor: 네비게이션 전달 인자 클래스명 및 패키지 변경
